requirements.txt
GoogleAppEngineCloudStorageClient==1.9.22.1
instagram-scraper==1.5.5
requests-toolbelt==0.8.0
